feat(example)!: move example specific logic out of `ElementSelector`.
There is now an `ElementSelectorDelegate` to manage the list of items to be displayed and an `ElementSelectorDelegateSubscription` to fire the insertion and removal animation independent of the buttons in the GUI having been pressed.

`Data` has been renamed to `ElementSelectorData`.

This will allow the `ElementSelector` widget to be moved into a seperate file or even library at some point, keeping the example clean!

BREAKING CHANGE: The `Element Selector` widget api changed a bit.