chore(plugin)!: delegate setting the icon image to new `TrayIconImageDelegate`
Users can now subclass the delegate to add their own sources (that are compatible with the available methods in the windows plugin).

Users can also store the delegate class and treat it as an abstract image.

BREAKING CHANGE: `image`, `picture` and `freeRessources` have been removed. Users have to pass a `ByteBuffer` now instead.